Extract artifact_state_projection.rs from turn.rs (parent #680)
Move the Phase-3 (Issue #659 Task 3.2) artifact-state projection helpers
out of `turn.rs` into a focused sibling module so `turn.rs` keeps
shrinking toward responsibility-focused units (parent meta-issue #680).

Four production helpers + three test seams extracted as free functions
taking `&mut Agent` / `&Agent` (matching `actor_loop_flow` /
`reply_retry` / earlier vertical-slice precedent):

- `task_contract_artifact_states` (production chokepoint, ledger
  authority + legacy shadow + divergence emit)
- `task_contract_artifact_states_legacy` (legacy derivation;
  side-effect seeder for Existing / Scaffold baseline ledger events)
- `task_contract_artifact_states_from_ledger` (ledger-projection
  derivation)
- `emit_artifact_state_projection_divergence` (masked observability
  emit; raw paths never leak, only bounded path-hash lists)
- 3 `#[cfg(test)]` test seams (for, _legacy_for, _from_ledger_for) used
  by the in-crate Phase 3 tests

Behavior unchanged: same iteration order over `required_artifacts`,
same `seed_artifact_ledger_*` side-effects in the legacy derivation,
same dedupe semantics, same masked `agent.artifact_ledger.divergence_detected`
emit shape (authority="ledger", projection name, bounded path-hash lists).

`pub(super)` limited / no facade re-export (DR3-001). Callers updated in
`turn.rs` (2 production self-callers) and
`artifact_ledger_phase3_tests.rs` (3 test seam call sites) to the
free-fn form.

LOC delta:
- turn.rs: 2,998 → 2,752 (-246)
- new module: +267

3,114 lib tests pass; `cargo fmt` / `cargo clippy --lib --all-targets
-- -D warnings` clean.
